requirements.txt
# ---- COMMON --- 
python-dotenv
fuzzywuzzy

# ---- API --- 
requests

# ---- SERVER --- 
fastapi
uvicorn
loguru
